backend/pyproject.toml
[tool.poetry]
name = "Recipe Browser"
version = "0.1.0"
description = ""
authors = ["helter daniel <helterdaniel@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dynaconf = "^3.1.12"
fastapi = "^0.94.0"
uvicorn = "^0.21.0"
alembic = "^1.10.2"
python-slugify = "^8.0.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
sqlalchemy = "^2.0.5.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyngrok = "^6.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
aiosqlite = "^0.18.0"
autoflake = "^1.4.0"
wemake-python-styleguide = "^0.17.0"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
